+++ src/tasks/CamelExportJBangTask.ts
@@ -22,9 +22,9 @@ import { CamelJBang } from "../requirements/CamelJBang";
 
 export class CamelExportJBangTask extends CamelJBangTask {
 
-	constructor(scope: WorkspaceFolder | TaskScope.Workspace, gav: string, runtime: string) {
+	constructor(scope: WorkspaceFolder | TaskScope.Workspace, gav: string, runtime: string, outputPath: string) {
 		super(scope,
 			'Create a Camel project',
-			new CamelJBang().createProject(gav, runtime));
+			new CamelJBang().createProject(gav, runtime, outputPath));
 	}
 }